Renamed landing page components for signed-out users; replace public components with signed-out versions and update item fetching logic
