+++ app/(public)/page.tsx
@@ -1,8 +1,9 @@
 import { SignedIn, SignedOut } from "@clerk/nextjs";
 import Footer from "@/components/Footer";
-import Hero from "@/features/landing-page/components/public/Hero";
-import Features from "@/features/landing-page/components/public/Features";
+import Hero from "@/features/landing-page/components/signed-out/Hero";
+import Features from "@/features/landing-page/components/signed-out/Features";
 import ListingsGrid from "@/features/landing-page/components/signed-in/ListingGrid";
+import { auth } from "@clerk/nextjs/server";
 import {
   getNewestItems,
   getMyItems,
@@ -10,7 +11,8 @@ import {
 
 export default async function Page() {
   const newestItems = await getNewestItems();
-  // const myItems = await getMyItems(userId as string);
+  const user = await auth();
+  const myItems = await getMyItems(user.userId as string);
   return (
     <main className="min-h-screen flex flex-col">
       <div className="flex-1 p-4">
@@ -30,7 +32,7 @@ export default async function Page() {
             />
             <ListingsGrid
               title="My Postings"
-              items={newestItems}
+              items={myItems}
               link="/postings"
               linkText="View all postings"
             />
